+++ lib/generators/surveyor/install_generator.rb
@@ -38,6 +38,8 @@ module Surveyor
       add_unique_index_on_access_code_and_version_in_surveys
       update_blank_versions_on_surveys
       api_ids_must_be_unique
+      create_survey_translations
+      add_input_mask_attributes_to_answer
     )
 
     def readme
@@ -67,6 +69,10 @@ module Surveyor
       end
     end
 
+    def routes
+      route('mount Surveyor::Engine => "/surveys", :as => "surveyor"')
+    end
+
     def assets
       if asset_pipeline_enabled?
         directory "app/assets"
@@ -84,6 +90,8 @@ module Surveyor
       copy_file "surveys/kitchen_sink_survey.rb"
       copy_file "surveys/quiz.rb"
       copy_file "surveys/date_survey.rb"
+      copy_file "surveys/languages.rb"
+      directory "surveys/translations"
     end
 
     def locales