+++ lib/surveyor/models/response_methods.rb
@@ -6,10 +6,7 @@ module Surveyor
         base.send :belongs_to, :response_set
         base.send :belongs_to, :question
         base.send :belongs_to, :answer
-        
-        # Scopes
-        base.send :default_scope, :order => "created_at ASC"
-        
+
         @@validations_already_included ||= nil
         unless @@validations_already_included
           # Validations
@@ -21,7 +18,7 @@ module Surveyor
 
         # Whitelisting attributes
         base.send :attr_accessible, :response_set, :question, :answer, :date_value, :time_value, :response_set_id, :question_id, :answer_id, :datetime_value, :integer_value, :float_value, :unit, :text_value, :string_value, :response_other, :response_group, :survey_section_id
-        
+
         # Class methods
         base.instance_eval do
           def applicable_attributes(attrs)
@@ -45,7 +42,7 @@ module Surveyor
       def default_args
         self.api_id ||= Surveyor::Common.generate_api_id
       end
-      
+
       def answer_id=(val)
         write_attribute :answer_id, (val.is_a?(Array) ? val.detect{|x| !x.to_s.blank?} : val)
       end
@@ -58,7 +55,12 @@ module Surveyor
       end
 
       def time_value=(val)
-        self.datetime_value = Time.zone.parse("#{Date.today.to_s} #{val}").to_datetime
+        self.datetime_value =
+          if val && time = Time.zone.parse("#{Date.today.to_s} #{val}")
+            time.to_datetime
+          else
+            nil
+          end
       end
 
       def date_value
@@ -66,7 +68,12 @@ module Surveyor
       end
 
       def date_value=(val)
-        self.datetime_value = Time.zone.parse(val).to_datetime
+        self.datetime_value =
+          if val && time = Time.zone.parse(val)
+            time.to_datetime
+          else
+            nil
+          end
       end
 
       def time_format
@@ -78,7 +85,23 @@ module Surveyor
       end
 
       def datetime_format
-        '%Y-%m-%d %H:%M'
+        '%Y-%m-%d %H:%M:%S'
+      end
+
+      def to_formatted_s
+        return "" if answer.nil? || answer.response_class.nil?
+        return case t = answer.response_class.to_sym
+               when :string, :text, :integer, :float
+                 send("#{t}_value".to_sym).to_s
+               when :date
+                 date_value
+               when :time
+                 time_value
+               when :datetime
+                 (read_attribute(:datetime_value).strftime( datetime_format ) unless read_attribute(:datetime_value).blank?) || ''
+               else
+                 to_s
+               end
       end
 
       def to_s # used in dependency_explanation_helper
@@ -99,7 +122,7 @@ module Surveyor
         }
 
         found = formats[answer.response_class]
-        found ?  datetime_value.strftime(found) : as(answer.response_class)
+        found ? datetime_value.try{|d| d.utc.strftime(found)} : as(answer.response_class)
       end
     end
   end