+++ lib/surveyor/models/survey_methods.rb
@@ -6,22 +6,23 @@ module Surveyor
     module SurveyMethods
       def self.included(base)
         # Associations
-        base.send :has_many, :sections, :class_name => "SurveySection", :order => 'display_order', :dependent => :destroy
-        base.send :has_many, :sections_with_questions, :include => :questions, :class_name => "SurveySection", :order => 'display_order'
+        base.send :has_many, :sections, :class_name => "SurveySection", :order => "#{SurveySection.quoted_table_name}.display_order", :dependent => :destroy
+        base.send :has_many, :sections_with_questions, :include => :questions, :class_name => "SurveySection", :order => "#{SurveySection.quoted_table_name}.display_order"
         base.send :has_many, :response_sets
+        base.send :has_many, :translations, :class_name => "SurveyTranslation"
 
         # Scopes
         base.send :scope, :with_sections, {:include => :sections}
-        
+
         @@validations_already_included ||= nil
         unless @@validations_already_included
           # Validations
           base.send :validates_presence_of, :title
           base.send :validates_uniqueness_of, :survey_version, :scope => :access_code, :message => "survey with matching access code and version already exists"
-          
+
           @@validations_already_included = true
         end
-        
+
         # Whitelisting attributes
         base.send :attr_accessible, :title, :description, :reference_identifier, :data_export_identifier, :common_namespace, :common_identifier, :css_url, :custom_class, :display_order
 
@@ -63,9 +64,17 @@ module Surveyor
         self.inactive_at = DateTime.now
         self.active_at = nil
       end
+
       def as_json(options = nil)
         template_paths = ActionController::Base.view_paths.collect(&:to_path)
-        Rabl.render(self, 'surveyor/export.json', :view_path => template_paths, :format => "hash")
+        Rabl.render(filtered_for_json, 'surveyor/export.json', :view_path => template_paths, :format => "hash")
+      end
+
+      ##
+      # A hook that allows the survey object to be modified before it is
+      # serialized by the #as_json method.
+      def filtered_for_json
+        self
       end
 
       def default_access_code
@@ -81,6 +90,13 @@ module Surveyor
         next_version = surveys.any? ? surveys.first.survey_version.to_i + 1 : 0
 
         self.survey_version = next_version
+      end
+
+      def translation(locale_symbol)
+        t = self.translations.where(:locale => locale_symbol.to_s).first
+        {:title => self.title, :description => self.description}.with_indifferent_access.merge(
+          t ? YAML.load(t.translation || "{}").with_indifferent_access : {}
+        )
       end
     end
   end